backend/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic-settings
numpy
scipy
aiortc
av
pytest
pytest-asyncio
pytest-anyio
httpx
